Clean up colorway support
